+++ src/test/java/wooteco/subway/service/LineServiceTest.java
@@ -5,9 +5,9 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
-import wooteco.subway.dao.LineRepository;
-import wooteco.subway.dao.LineRepositoryImpl;
-import wooteco.subway.dao.StationRepository;
+import wooteco.subway.domain.repository.LineRepository;
+import wooteco.subway.domain.repository.LineRepositoryImpl;
+import wooteco.subway.domain.repository.StationRepository;
 import wooteco.subway.domain.Line;
 import wooteco.subway.dto.LineRequest;
 import wooteco.subway.dto.LineResponse;
