+++ src/test/java/wooteco/subway/service/StationServiceTest.java
@@ -5,8 +5,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
-import wooteco.subway.dao.StationRepository;
-import wooteco.subway.dao.StationRepositoryImpl;
+import wooteco.subway.domain.repository.StationRepository;
+import wooteco.subway.domain.repository.StationRepositoryImpl;
 import wooteco.subway.domain.Station;
 import wooteco.subway.dto.StationRequest;
 import wooteco.subway.dto.StationResponse;
